Fixed write concern for lock record with Mongoid5 and refactored wrapper.
