+++ .rubocop_todo.yml
@@ -1,20 +1,20 @@
 # This configuration was generated by `rubocop --auto-gen-config`
-# on 2015-09-25 09:10:09 -0400 using RuboCop version 0.29.1.
+# on 2015-10-21 11:03:44 -0400 using RuboCop version 0.29.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 2
+# Offense count: 1
 Metrics/AbcSize:
-  Max: 19
+  Max: 18
 
-# Offense count: 25
+# Offense count: 37
 # Configuration parameters: AllowURI, URISchemes.
 Metrics/LineLength:
   Max: 184
 
-# Offense count: 4
+# Offense count: 3
 # Configuration parameters: CountComments.
 Metrics/MethodLength:
   Max: 25
